+++ projects/ods-loading-indicator/src/lib/ods-loading-indicator.module.ts
@@ -1,0 +1,17 @@
+import {NgModule} from "@angular/core";
+import {OdsLoadingIndicatorComponent} from "./ods-loading-indicator.component";
+
+@NgModule(
+{
+	declarations:
+	[
+		OdsLoadingIndicatorComponent
+	],
+	imports:[],
+	exports:
+	[
+		OdsLoadingIndicatorComponent
+	]
+})
+
+export class OdsLoadingIndicatorModule{}