+++ projects/ods-loading-indicator/src/public-api.ts
@@ -1,0 +1,6 @@
+/*
+ * Public API Surface of ods-loading-indicator
+ */
+
+export * from "./lib/ods-loading-indicator.component";
+export * from "./lib/ods-loading-indicator.module";